requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP Client
aiohttp==3.9.1
requests==2.31.0

# Browser Automation
playwright==1.41.0

# AI/LLM
anthropic==0.17.0

# Data Processing
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2  # For Excel files
xlrd==2.0.1  # For older Excel files

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Image Processing
Pillow==10.2.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.1

# Utilities
python-multipart==0.0.6  # For form data
